pyproject.toml
[project]
name = "devx"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [{ name = "connormaglynn" }]
requires-python = ">=3.14"
dependencies = [
    "pygithub==2.3.0",
    "typer>=0.24.1",
]

[project.scripts]
devx = "devx.main:app"

[build-system]
requires = ["uv_build>=0.10.8,<0.11.0"]
build-backend = "uv_build"
